@@ -12,11 +12,13 @@
   "author": "Emily Turner",
   "license": "ISC",
   "dependencies": {
-    "react": "^0.14.3",
-    "react-dom": "^0.14.3",
+    "body-parser": "^1.15.2",
+    "express": "^4.14.0",
+    "object-assign": "^2.0.0",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-router": "^4.1.1",
-    "style-loader": "^0.14.1",
-    "object-assign": "^2.0.0"
+    "style-loader": "^0.14.1"
   },
   "devDependencies": {
     "babel-core": "^6.3.13",
