@@ -24,7 +24,7 @@ app.use(express.static("./public"));
 // });
 
 app.get('/', function(req, res) {
-  res.sendFile(__dirname + '/public/index.html');
+  res.sendFile(path.join(__dirname + '/public/index.html'));
 });
 
 // Listener
